Extract `Fluent::Test::StartupShutdown`
