+++ lib/fluent/test.rb
@@ -23,6 +23,7 @@ require 'fluent/test/output_test'
 require 'fluent/test/filter_test'
 require 'fluent/test/parser_test'
 require 'fluent/test/formatter_test'
+require 'fluent/test/startup_shutdown'
 require 'serverengine'
